ci(dashboard): run the dashboard tests
The `dashboard` job ran lint, format and build, and never invoked jest. Nothing
in it rendered a component, which is how a crash on every filter interaction on
the managed-jobs page reached master with this job green (fixed in #10459), and
how both existing suites came to fail without anyone noticing.

Turning the job on required fixing them, not skipping them:

- `data/connectors/jobs.test.jsx` could not even load: `MANAGED_JOBS_SUMMARY_ARGS`
  is a `const` exported from a module that sits on an import cycle (jobs ->
  plugins/dataEnhancement -> plugins/PluginProvider -> lib/cache-preloader ->
  jobs), so the re-entrant importer read it inside its temporal dead zone. The
  functions crossing that cycle are hoisted declarations and were unaffected;
  this value was not. It is a frozen config object with no dependencies of its
  own, so it moves to `data/connectors/constants`, which imports nothing. One
  caller was already working around this with a dynamic import; that goes away.

- `elements/version-display.test.jsx` asserted on tooltip content while the
  tooltip was closed. Its NextUI host mounts content only when open, so every
  case failed from the day they were written -- including at the commit that
  added them. The content moves into an exported `VersionTooltipContent`, which
  is the part with the behaviour worth asserting on, and the test targets that.
  One expectation was also backwards: with no plugins the label is `Commit`, and
  `Core commit` only earns its qualifier when plugin commits sit beside it.

`npx jest` -> 16 suites, 165 tests, all passing.
